feat: create nex email template
